[스무디] 모델 변경
- User 모델 생성
- 각 모델에 생성/수정날짜 추가

+++ 2019/Spring(boot) study/smoothie/src/main/java/com/example/smoothie/controller/CommentController.java
@@ -20,7 +20,7 @@ public class CommentController {
     }
 
     @GetMapping("/{id}")
-    public Comment get(@PathVariable Integer id) {
+    public Comment get(@PathVariable Long id) {
         return service.get(id);
     }
 
@@ -35,7 +35,7 @@ public class CommentController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public void delete(@PathVariable Long id) {
         service.delete(id);
     }
 }
